Create new navigator called RootNavigator
