+++ App.tsx
@@ -1,12 +1,6 @@
 import React, {useEffect} from 'react';
-import {NavigationContainer} from '@react-navigation/native';
-import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import TabNavigator from './src/navigators/TabNavigator';
-import {DetailsScreen} from './src/screens/DetailsScreen';
-import {PaymentScreen} from './src/screens/PaymentScreen';
 import SplashScreen from 'react-native-splash-screen';
-
-const stack = createNativeStackNavigator();
+import { RootNavigator } from './src/navigators/RootNavigator';
 
 const App = () => {
   useEffect(() => {
@@ -14,27 +8,7 @@ const App = () => {
   }, []);
 
   return (
-    <NavigationContainer>
-      <stack.Navigator
-        initialRouteName="Tab"
-        screenOptions={{headerShown: false}}>
-        <stack.Screen
-          name="Tab"
-          component={TabNavigator}
-          options={{animation: 'slide_from_bottom'}}
-        />
-        <stack.Screen
-          name="Details"
-          component={DetailsScreen}
-          options={{animation: 'slide_from_bottom'}}
-        />
-        <stack.Screen
-          name="Payment"
-          component={PaymentScreen}
-          options={{animation: 'slide_from_bottom'}}
-        />
-      </stack.Navigator>
-    </NavigationContainer>
+    <RootNavigator />
   );
 };
