+++ src/navigators/RootNavigator.tsx
@@ -1,0 +1,34 @@
+import React from 'react';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import { NavigationContainer } from '@react-navigation/native';
+import TabNavigator from '../navigators/TabNavigator';
+import {DetailsScreen} from '../screens/DetailsScreen';
+import {PaymentScreen} from '../screens/PaymentScreen';
+
+const stack = createNativeStackNavigator();
+
+export const RootNavigator: React.FC = () => {
+  return (
+    <NavigationContainer>
+      <stack.Navigator
+        initialRouteName="Tab"
+        screenOptions={{headerShown: false}}>
+        <stack.Screen
+          name="Tab"
+          component={TabNavigator}
+          options={{animation: 'slide_from_bottom'}}
+        />
+        <stack.Screen
+          name="Details"
+          component={DetailsScreen}
+          options={{animation: 'slide_from_bottom'}}
+        />
+        <stack.Screen
+          name="Payment"
+          component={PaymentScreen}
+          options={{animation: 'slide_from_bottom'}}
+        />
+      </stack.Navigator>
+    </NavigationContainer>
+  );
+};
